pyproject.toml
[project]
name = "javis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiomysql>=0.2.0",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.3",
    "duckduckgo-search>=8.0.0",
    "httpx>=0.28.1",
    "logfire>=3.12.0",
    "mcp[cli]>=1.6.0",
    "pydantic-ai-slim[mcp]>=0.0.53",
    "python-telegram-bot>=22.0",
    "redis>=5.2.1",
]
